+++ backend/src/main/java/woowacourse/touroot/travelogue/day/dto/TravelogueDayResponse.java
@@ -1,0 +1,7 @@
+package woowacourse.touroot.travelogue.day.dto;
+
+import java.util.List;
+import woowacourse.touroot.travelogue.place.dto.TraveloguePlaceResponse;
+
+public record TravelogueDayResponse(List<TraveloguePlaceResponse> places) {
+}
