+++ backend/src/main/java/woowacourse/touroot/travelogue/day/repository/TravelogueDayRepository.java
@@ -1,7 +1,7 @@
-package woowacourse.touroot.travelogueday.repository;
+package woowacourse.touroot.travelogue.day.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import woowacourse.touroot.travelogueday.domain.TravelogueDay;
+import woowacourse.touroot.travelogue.day.domain.TravelogueDay;
 
 public interface TravelogueDayRepository extends JpaRepository<TravelogueDay, Long> {
 }
