+++ backend/src/main/java/woowacourse/touroot/travelogue/photo/repository/TraveloguePhotoRepository.java
@@ -1,7 +1,7 @@
-package woowacourse.touroot.traveloguephoto.repository;
+package woowacourse.touroot.travelogue.photo.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import woowacourse.touroot.traveloguephoto.domain.TraveloguePhoto;
+import woowacourse.touroot.travelogue.photo.domain.TraveloguePhoto;
 
 public interface TraveloguePhotoRepository extends JpaRepository<TraveloguePhoto, Long> {
 }
