+++ backend/src/main/java/woowacourse/touroot/travelogue/place/repsitory/TraveloguePlaceRepository.java
@@ -1,7 +1,7 @@
-package woowacourse.touroot.travelogueplace.repsitory;
+package woowacourse.touroot.travelogue.place.repsitory;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import woowacourse.touroot.travelogueplace.domain.TraveloguePlace;
+import woowacourse.touroot.travelogue.place.domain.TraveloguePlace;
 
 public interface TraveloguePlaceRepository extends JpaRepository<TraveloguePlace, Long> {
 }
