--- backend/src/main/java/woowacourse/touroot/travelogueday/dto/TravelogueDayResponse.java
@@ -1,7 +1,0 @@
-package woowacourse.touroot.travelogueday.dto;
-
-import java.util.List;
-import woowacourse.touroot.travelogueplace.dto.TraveloguePlaceResponse;
-
-public record TravelogueDayResponse(List<TraveloguePlaceResponse> places) {
-}
